♻️ Meilleur découpage des permission
