+++ packages/domain/utilisateur/src/permission/role/caisseDesDépôts.ts
@@ -1,0 +1,11 @@
+import { Policy } from '..';
+
+export const permissionCaisseDesDépôts: Policy[] = [
+  // Garanties financières
+  'garantiesFinancières.actuelles.consulter',
+  'garantiesFinancières.dépôt.consulter',
+  'garantiesFinancières.mainlevée.lister',
+
+  // Achèvement
+  'achèvement.consulter',
+];
