+++ packages/domain/utilisateur/src/permission/role/grd.ts
@@ -1,0 +1,10 @@
+import { Policy } from '..';
+
+export const permissionGRD: Policy[] = [
+  'réseau.gestionnaire.consulter',
+  'réseau.raccordement.consulter',
+  'réseau.raccordement.listerDossierRaccordement',
+  'réseau.raccordement.date-mise-en-service.transmettre',
+  'réseau.raccordement.date-mise-en-service.modifier',
+  'réseau.raccordement.référence-dossier.modifier',
+];
