+++ packages/domain/utilisateur/src/permission/role/index.ts
@@ -1,0 +1,7 @@
+export { permissionAdmin, permissionDgecValidateur } from './admin';
+export { permissionDreal } from './dreal';
+export { permissionPorteurProjet } from './porteurProjet';
+export { permissionAcheteurObligé } from './acheteurObligé';
+export { permissionCRE } from './cre';
+export { permissionCaisseDesDépôts } from './caisseDesDépôts';
+export { permissionGRD } from './grd';
